apache-flask/app/requirements.txt
mysql-connector-python
secure-smtplib
Flask
Flask-APScheduler
pandas
